Update: integro repo con contenidos del repo firmware_v3

+++ libs/sapi/sapi_v0.5.2/soc/peripherals/inc/sapi_spi.h
@@ -55,7 +55,7 @@ bool_t spiInit( spiMap_t spi );
 
 bool_t spiRead( spiMap_t spi, uint8_t* buffer, uint32_t bufferSize );
 
-bool_t spiWrite( spiMap_t spi, uint8_t* buffer, uint32_t bufferSize);
+bool_t spiWrite( spiMap_t spi, uint16_t* buffer, uint32_t bufferSize);
 
 /*==================[c++]====================================================*/
 #ifdef __cplusplus
